requirements.txt
pytelegrambotapi==4.19.1
pillow==10.1.0
pillow-heif==1.1.1
opencv-python==4.9.0.80
numpy==1.24.3